added basic MongoDB support into a-frame. the abstracting class is nowhere near the power of the SQL class but it doesn't need to be since the PHP Mongo object has just about all the needed features and simplicity...to abstract away would only remove features.
now, there is a database type AFRAME_DB_MODE_MONGODB which HAS to be set in the $config['db']['dsn'] array as mode (eg $config['db']['dsn'] = array(... 'mode' => AFRAME_DB_MODE_MONGODB', ...)) for the mongo db driver to be used. this makes it bacwards compatible with the mysql driver.

one main problem with the current setup is that it doesn't support using multiple databases at once. this will most likely be rememdied once we actually do a project that requires this. until, its *SQL or Mongo. deal with it.

note that all of this took some refactoring. the base_model was slpit up so that all database code actually lives in a base_db object under classes/base/db/. the classes in there are ALL named base_db, but included dynamically based on what type of database is selected. this will make programming in multiple database types a tad uncomfortable in the future, but for now it doesn't matter at all. also, the original classes/db.php file is now classes/db/sql.php, and a new one for mongo was added under classes/db/mongodb.php.

so far so good, but won't merge into master until it's solid as a rock.